+++ src/parser/parse_utils1.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   parse_utils1.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: nmihaile <nmihaile@student.42heilbronn.    +#+  +:+       +#+        */
+/*   By: bwerner <bwerner@student.42heilbronn.de>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/08 09:56:53 by nmihaile          #+#    #+#             */
-/*   Updated: 2024/08/09 16:11:34 by nmihaile         ###   ########.fr       */
+/*   Updated: 2024/08/10 20:14:58 by bwerner          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -49,7 +49,7 @@ void	ft_skipspace(char **str)
 // sign_dpoint_dplaces[2] = 0;
 // defaults are:
 // ft_atod(&str, 0, int[3]{1, -1, 0})
-double ft_atod(char **str, double nbr, int sign_dpoint_dplaces[3])
+float ft_atod(char **str, float nbr, int sign_dpoint_dplaces[3])
 {
 	ft_skipspace(str);
 	sign_dpoint_dplaces[0] = get_sign(str);
